Login Pg Separate & set up Jesus
Changes Made:
errors are not prompting through snackbar anymore. More conventional error messages. Back button has been removed after logging in. I need to add more validation parameters before actual login.
